Flickering, no anymore
Okay, so now we started factoring the code. Right now we only have the wheel on display, and it's not flickering. It's able to bring up whatever sect has been picked, segment, whatever, right? The last part of the wheel that hit the pointer, it's able to bring that up. And yeah, I've removed the other UI elements, that's the two panels, and also those jars at the bottom, and the buttons at the top, I've removed them for now, but everything seems to work accordingly. It's MVC driven, and it started flickering.