Failing to generate sample population for SRS
Okay, so now the only problem that we're meeting right from the above is we're failing to populate the sample array. For some reason, I think the condition that we have is actually creating a sample population that is bigger than the sample size. So far, everything else has worked, but now we just have to figure out this final bit.